refactor:  liquor앱 게시글 추가, 수정, 삭제 기능 fix
- 회원도 게시글 추가, 수정, 삭제 기능 사용 가능한 문제 발견
- models.py에 관리자인지 확인하는 코드 추가

+++ liquor/views.py
@@ -10,7 +10,7 @@ from .serializers import LiquorListSerializer, LiquorDetailSerializer
 
 # 주류 목록 조회(GET/ 누구나 이용 가능) 및 등록(POST/ 관리자만 가능)
 class LiquorListView(APIView):
-    permission_classes = [IsAuthenticatedOrReadOnly]  # 관리자만 가능 or 누구나 이용 가능
+    permission_classes = [IsAuthenticatedOrReadOnly]  # 인증된 회원(회원or관리자)만 가능 or 누구나 이용 가능
 
     # 주류 목록 조회
     def get(self, request):
@@ -20,9 +20,10 @@ class LiquorListView(APIView):
 
     # 게시글 등록
     def post(self, request):
-        # self.permission_classes = [IsAdminUser]  # 관리자만 등록 가능
-        # self.check_object_permissions(request)  # 관리자가 맞는지 체크
-
+        #관리자인지 확인하는 코드 추가
+        if not request.user.is_superuser:
+            return Response({"detail": "접근 불가 / 관리자만 가능"}, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = LiquorDetailSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -48,7 +49,7 @@ class LiquorBookmarkView(APIView):
 
 # 주류 디테일 페이지 조회(GET/누구나 이용 가능) 및 수정(PUT/관리자만), 삭제(DELETE/관리자만)
 class LiquorDetailView(APIView):
-    permission_classes = [IsAuthenticatedOrReadOnly]  # 관리자만 가능 or 누구나 이용 가능
+    permission_classes = [IsAuthenticatedOrReadOnly]  # 인증된 회원(회원or관리자)만 가능 or 누구나 이용 가능
 
     # 주류 디테일 페이지 조회
     def get(self, request, pk):
@@ -58,9 +59,11 @@ class LiquorDetailView(APIView):
 
     # 게시글 수정
     def put(self, request, pk):
-        # self.permission_classes = [IsAdminUser]  # 관리자만 등록 가능
-        # self.check_object_permissions(request)  # 관리자가 맞는지 체크
         liquor = get_object_or_404(Liquor, pk=pk)
+        #관리자인지 확인하는 코드 추가
+        if not request.user.is_superuser:
+            return Response({"detail": "접근 불가 / 관리자만 가능"}, status=status.HTTP_403_FORBIDDEN)
+        
 
         serializer = LiquorDetailSerializer(liquor, data=request.data, partial=True)  # partial은 부분 수정
         if serializer.is_valid():
@@ -70,9 +73,11 @@ class LiquorDetailView(APIView):
 
     # 게시글 삭제
     def delete(self, request, pk):
-        # self.permission_classes = [IsAdminUser]  # 관리자만 등록 가능
-        # self.check_object_permissions(request)  # 관리자가 맞는지 체크
         liquor = get_object_or_404(Liquor, pk=pk)
+        #관리자인지 확인하는 코드 추가
+        if not request.user.is_superuser:
+            return Response({"detail": "접근 불가 / 관리자만 가능"}, status=status.HTTP_403_FORBIDDEN)
+        
 
         liquor.delete()
         return Response({"message": "게시글 삭제 완료"}, status=status.HTTP_403_FORBIDDEN)
